Add test for vocabulary ast storage and move common test logic to separate file
